requirements.txt
pydantic~=1.9.2
python-dotenv~=0.20.0
requests~=2.28.1